Fix the bug "Incoming call from (null)" (Alban Crequy)
The status icon uses empathy_contact_get_name() to print the name of the
calling contact. It uses empathy_contact_run_until_ready() to wait the
return of GetAliases(). If GetAliases() failed for whatever reason, the
name will be NULL and empathy_contact_get_name() will fallback to the
id. If GetHandle() has not yet returned, the id is not ready and
empathy_contact_run_until_ready() must wait the id to be available.

svn path=/trunk/; revision=1179