Only do sounds/notifications for text messages when they involve the user
When in a chatroom notifications and sounds should only happen when the users
is part of the message (either someone is speaking to them or referring to
them), otherwise the user will go crazy when in a lot of different muc
channels. For a peer to peer text channel, all messages sent are targetted at
the user so always do sounds/notifications in that case.