Break after deleting current item in chatroom list
When I signed off, Empathy exploded. l was 0x33, which oddly enough was
secretly not the address of a GList. :-) I think what happened is that
empathy_chatroom_manager_remove() deleted the list node currently
pointed to by l; then the loop advanced to l->next, which was now
garbage.

We don't need to keep iterating once we find the chatroom we were
looking for, so break out of the loop.