Correctly propagate Provide/Accept errors
Previously, if myerr was set as well as error, myerr was cleared, but
not set to a copy of error. So then the function continued as normal,
and crashed.

Also, having checked if myerr is set, we should probably log the message
from that, rather than from the (possibly NULL) error. :)